hackathon/requirements.txt
aioredis==1.2.0
asgiref==3.1.4
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.7.1
Automat==0.7.0
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
Django==2.2.3
django-betterforms==1.2
djangorestframework==3.9.4
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
msgpack==0.6.1
Pillow==6.1.0
pycparser==2.19
PyHamcrest==1.9.0
pypiwin32==223
pytz==2019.1
pywin32==224
six==1.12.0
sqlparse==0.3.0
Twisted==19.2.1
txaio==18.8.1
zope.interface==4.6.0
